chg: makefile; test方式
