+++ .github/workflows/tests.yml
@@ -47,7 +47,7 @@ jobs:
       # 进行测试
       - name: Test
         run: |
-          make alltests
+          make test
         env:
           MTFMT_BUILD_USE_LTO: "1"
           MTFMT_BUILD_C_DEFS: "-D_MSTR_USE_MALLOC=${{ matrix.buildin-alloc }} -D_MSTR_USE_HARDWARE_DIV=${{ matrix.buildin-div }}"
