+++ tests/test_fmt_sty.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: LGPL-3.0
 /**
- * @file    fmt_sty.c
+ * @file    test_fmt_sty.c
  * @author  向阳 (hinata.hoshino@foxmail.com)
  * @brief   值格式化样式的测试
  * @version 1.0
@@ -10,28 +10,12 @@
  *
  */
 #include "helper.h"
+#include "main.h"
 #include "mm_fmt.h"
 #include "mm_heap.h"
 #include "unity.h"
 #include <stddef.h>
 #include <stdio.h>
-
-#define RUNTIME_HEAP_SIZE 2048
-
-/**
- * @brief 堆
- *
- */
-static byte_t heap[RUNTIME_HEAP_SIZE];
-
-void setUp(void)
-{
-    mstr_heap_init(heap, RUNTIME_HEAP_SIZE);
-}
-
-void tearDown(void)
-{
-}
 
 void fmt_sty_bin(void)
 {
@@ -76,15 +60,4 @@ void fmt_sty_hex_prefix(void)
     EVAL(mstr_format("@{0:i32:x}@{0:i32:X}@", &s, 1, 0xa5));
     ASSERT_EQUAL_STRING(&s, "@0xa5@0XA5@");
     mstr_free(&s);
-}
-
-int main()
-{
-    UNITY_BEGIN();
-    RUN_TEST(fmt_sty_bin);
-    RUN_TEST(fmt_sty_oct);
-    RUN_TEST(fmt_sty_dec);
-    RUN_TEST(fmt_sty_hex);
-    RUN_TEST(fmt_sty_hex_prefix);
-    return UNITY_END();
 }
